+++ LearningTrackerProject/LearningTracker/NotesManager.cs
@@ -1,0 +1,3 @@
+namespace LearningTracker;
+
+public class NotesManager{}
