+++ LearningTrackerProject/LearningTracker/ProgressManager.cs
@@ -1,0 +1,3 @@
+namespace LearningTracker;
+
+public class ProgressManager{}
